Make dForm know how to get the fieldVal from the bizobj and return it to the control. Make dControl know how to ask the form for the value and refresh itself.
Update testAll in ui/uiwx/classes/test to add a 
bizobj (the bizZip) to the test form. Now when
testAll is run, textbox values are updated when
the vcr is used for navigation.

Next up is probably to implement save and cancel
functionality at the dForm level. Also, the dControls
are notifying the dForm when the user changed the
value, but the dForm isn't notifying the bizobj
yet.

I keep starting other work and then coming back to
Dabo. Not very good on the pocketbook... :)