Begin work on dGrid: not working at this point.
Add dMessageBox.stop() for displaying messages
that communicate 'cannot continue' responses
to the user.

Add a stop() message to dForm.save() when the 
bizobj communicates that the save can't happen,
perhaps because of a biz rule violation.

Add dForm.activeControlValid() which has the
effect of updating the bizobj with the control's
current value if necessary, say before a 'save'
operation. Otherwise, if the user changed a 
field value and clicked 'save' without first
setting focus to a different control, their 
changes to that field would not be included 
in the commit.