Added 'seek' functionality to the bizobj and cursor classes. Behavior is patterned on the VFP SEEK() command.
Modified dBizobj's __getattr__ and setattr__ so that you can now refer to field values as if they were attributes of the bizobj, and setting these attributes will automatically update the cursor. This will allow a great deal of simplification and readability in bizobj code, since the code will no longer need to be cluttered with 'getFieldVal()' and 'setFieldVal()' calls throughout.