Combined Ted's delete icon with the 'new' icon to get a delete icon that seems to fit in.
Removed the help provider from the app object because we 
aren't using it (yet) and it got moved to the main wx module
in wxPython 2.5. At some point we'll move to 2.5 and reenable
it.

In 2.5, image lists must contain images of exactly the same size,
something that wasn't enforced in 2.4. Resized a few images to
conform. Also, by default image lists start with one image in the
list, not zero. This caused a segfault in 2.5.