Started implementing the select page.
While working on that, I discovered that in 2.5 the 
setfocus/killfocus code wasn't running in the controls
on the edit page, and remembered that Ed mentioned 
something like that the other day. The common factor
must be GTK2 versus GTK1, and it came down to adding
event.Skip() in the form's SetFocus code - I had just
added SetFocus/KillFocus code in the form on my airplane
ride. 

Tested this in wx2.4/GTK1 as well as in wx2.5/GTK2.

It sure is hairy how my changes may work fine on 
my computer yet break something on another computer. I
guess that is what we get for developing multi-platform
using the native toolkits on each platform.

By the way, event binding will be much simpler in 2.5,
as they've introduced the form:

   self.Bind(wx.EVT_SET_FOCUS, self.OnSetFocus)

to bind events to methods, which feels a lot like VFP's
BINDEVENT() does it not?