Add NoRecordsError and QueryError, and raise those exceptions from dCursorMixin instead of the generic dError. Rework dForm to import dError into a separate namespace instead of into the global namespace. I think this should be done for dCursorMixin as well. Rework dForm to act appropriately upon these specific exceptions.
Fix some minor things in dGrid - problems would occur if there weren't any 
records in the dataset.

Make the select page make select option fields based on the columnDefs dict
in dFormDataNav. This is the same source of columns for grids and fields to
display in the edit page. Re-route dFormDataNav's requery() function to call
the select page's requery(), which sets the where clause of the SQL and then
ultimately calls dForm.requery(). Currently, this SQL gets manipulated in 
dFormDataNav's sqlBuilder, but I think ultimately this sqlBuilder needs to 
be mixed-in to either dBizobj or to dCursor - somewhere 'lower down' than
the UI. 

Update dDataControlMixin to recognize that there isn't a current record, as
when there are no records in the dataset, and to disable itself and set its
value to blank. Conversely, when it recognizes that there is current data
it will enable itself.