Add dError.BusinessRuleViolation exception.
Ran into issue with adding a new record in recipes, and it turns out to
be because I coded dSqlBuilderMixin to format the sql nicely with newlines
but this regex in dCursorMixin.__setStructure():
        pat = re.compile("(\s*select\s*.*\s*from\s*.*\s*)((?:where\s(.*))+)\s*", re.I | re.M)
doesn't account for the newlines, so it doesn't find the where clause. Rather 
than figure out that regex, it was easier just to take the newlines out of 
dSqlBuilderMixin for now.

In dCursorMixin.new(), we were catching all exceptions and then raising
dError, which inappropriately hides unexpected exceptions. Better just to
take the try block out and let the unexpected exceptions propagate up. Until
I did this, I wasn't seeing the cause of my problem (first, it was that 
global name 'sql' isn't defined in __setStructure, and then it was a MySQL
exception having to do with bad SQL because the second where clause: 'where
1=0' was tagged onto the end.

Add 'value' and 'stringMatch' selectTypes to the select page auto generation
of controls.

Fix problem in dGrid that I introduced recently where after the row number in
the bizobj changes the grid's currently highlighted row wasn't changing.