Remove extraneous getSQL() from the bizobj, now that getSQL() is a passthrough to the cursor's sqlbuilder.
I found another example in dCursorMixin where we catch all exceptions
and then raise dError. This isn't right, as we want the actual exception
to propagate up to the global error handler, not to hide it.

Remove sqlBuilder from dFormDataNav, as it is now exposed in the bizobj.
Set up the requery function in the select page to interact correctly.

Remove all the old stuff from my prior incarnation of Dabo, as it is all
pretty much reimplemented now on a much stronger platform.

Introduction of dPemMixin, which provides a way to define properties of
dControls and map them to wx behavior. For instance, you can now set the
size of a text box by calling:
    textBox.Left = 12
    textBox.Top = 20
Not only will this make it easier to provide other UI interfaces (qt, etc.)
while still keeping the property names consistent, but it should also make
it easier to define subclasses of the dControls, once we have the resource
format all worked out. Also, VFP'ers will feel right at home using familiar
property names instead of having to remember wx's function syntax every time.

Make dControlMixin and dDataControlMixin inherit from dPemMixin.

Make dDataControlMixin define a few extra properties, as these are items
that our users are highly likely to want to interact with (DataSource,
DataField, and SelectOnEntry).

With this commit, even more of Dabo is dependent on wxPython 2.5, as I've
used Python new-style classes in the mixins and wxPython 2.4 will fail
when trying to use new-style mixins.