Set more property definitions in controls and forms, matching up wx get/set functions to VFP property naming styles. One good side effect is that you can now always refer to self.Parent to get the containing object. As soon as only dControls are in use everywhere, we'll be able to say 'self.Parent.Parent...'. Next, I'll probably add a self.Form property which would be analagous to VFP's ThisForm.
Make messagebox's center themselves on top of the active form, not on 
top of the top level form.

Make the main menu bar appear on all forms, which is more Linux-like. This
should have no effect on Windows where there is just one main menu bar on the
parent MDI form but I have yet to test that.