There is an inconsistency in dBizobj.isChanged() versus dCursorMixin.isChanged(). dBizobj's docstring states that it is going to just check the current record, while dCursorMixin's docstring states it will check all records, which is what it does.
I didn't change the docstrings, but I did add an allRows argument to both of these
functions, along with logic in dCursorMixin to act appropriately based on the 
parameter. I made the default value of these arguments match the docstrings, so
allRows=False in dBizobj, and allRow=True in dCursorMixin.

Changed the name of dLocalize.loc() to dLocalize._() to match the il8n conventions,
but loc() still works as well. Changed dForm's loc() calls to _() calls.

Added a 'save changes' prompt to dForm.requery() when there are unsaved changes,
allowing Yes/No/Cancel choices.