Make dLabel not auto-resize to the length of the text. Fix a couple minor bugs.
Add the following properties, most of which are saved in the WindowStyleFlags
and so otherwise would be kind of difficult to use:

dCheckBox:
    .Alignment (left, right)
    
dEditBox:
    .Alignment (left, center, right)
    .ReadOnly
    
dLabel:
    .Alignment (left, center, right)
    
dPageFrame:
    .TabPosition (top, left, right, bottom)

dSpinner:
    .SpinnerWrap (does the value wrap from low to high, and vice-versa)
    .SpinnerArrowKeys (can the user spin up/down with the keyboard)
    
dTextBox:
    .Alignment (left, center, right)
    .ReadOnly (like Enabled=False but allows text selection)
    .PasswordEntry (if True, echoes asterisks instead of Value)

All forms:
    .BorderResizable
    .ShowMaxButton
    .ShowMinButton
    .ShowCloseButton
    .ShowCaption
    .ShowSystemMenu
    .TinyTitleBar
    
All controls and forms:
    .BorderStyle (none, simple, raised, sunken)
    .WindowHandle (platform-specific Hwnd)

Note that in most cases I haven't tested any property interactions yet, but 
in theory they should all work. Some of these won't work on live objects, so
in the future when integrating with the designer we'll probably have to recreate
the object after some of these properties change. My feeling is that most testing
of the properties can wait until the designer is instantiating these controls, at
which point we'll have a convenient interactive testing environment.

Note also that I chose to skip over some possibilities, such as making a 3-state
checkbox control and some other styles that only work on this or that platform.
We can revisit these in the future and choose to put them in, but my feeling is
that if it doesn't apply to all platforms, it shouldn't apply to any platform.
Keeping Dabo at the lowest common functionality will reduce surprises and headaches
for our users later, when they try to port their app from Windows to Mac.