Fix some import statements. Fix bug in property 'ToolTipText'. Fix ui test script to use the new property interface.
Rework the dControl's constructors to pretty much conform to wx's constructor
format. There are some exceptions, and I haven't done the dForm's yet. To be
completely safe, anything beyond parent and id should use named arguments even
though theoretically positional arguments should work.

Add a requery button to the select page of dFormDataNav, and try to set it as
the default button (click on <Enter>) but at least on Linux that isn't working
as advertised. Rework the select page to use dControls instead of wx controls
directly and use the property interface.