Change some of my " to ' <g>.  Localize some strings.
Change multi-state properties, such as Alignment, to have string
identifiers instead of integer identifiers.

Give dForm the wx interface, just like I did for the dControls earlier.
Remove setupResources() from dForm as I don't think that is the interface
we'll use (I don't think we'll use resources at all).

Fix some properties to work correctly. This was after testing with 
dTextBox in daboDesigner. I need to figure out how to deal with the 
styles that have to be passed to the constructor, not altered after the
object has already been created (Alignment, for example). I'm thinking
that we'll have to get a copy of the object's properties, destroy it, 
and then recreate it.