Separate out the property initialization, based on whether the property needs to be passed as a style parameter to the constructor (property will be read-only at runtime), or whether the property must be set after the object is fully instantiated (property will be read/write at runtime unless specified differently in the property assignment).
Add getPropertyInfo() to dPemMixin, which returns a dict of useful information
on the passed property name, including the data type, the editor to use, and 
the docstring. This is to make the designer much more generic: it can just
query the object for most of the information it needs.

Part of the logic of the above involves checking for the editor to use by
calling a special method which is named similarly to the given property.
If that method doesn't exist, a default value for editor info is returned, 
based on data type, which is how I've left it for most of the properties.
The only properties I've specifically defined editors for so far are the ones 
that have a set number of choices to be presented in a list, such as 'Alignment'.

Add property 'AutoResize' to dLabel.