Increase robustness of property setters by coercing proper data types and raising exceptions with improper values.
Fix a few places where my strategy for setting/deleting window style
flags wasn't up to snuff.

Add 'showInDesigner' key to the getPropertyInfo dict. This tells the 
designer whether to show the property at all. Set WindowHandle, Size,
and Position to not show in the designer.

Add 'preInitProperty' key to the getPropertyInfo dict. This tells the
designer where to put this property setting in the class's source code.
Some properties need to be set before the object is fully constructed,
while most need to be set after the object is fully constructed. Set 
the few properties I know about that need to be defined early.

Add some editorInfo overrides to some properties, where the defaults 
wouldn't be adequate. For instance, the default for integers is -65536
to 65536, which is completely inappropriate for properties like Width
and Height, which can't go below 0.