Fix some bugs in dOptionGroup.
Experimental: I'm actually scared at how easy this was to implement, but I added
a reCreate() function to dPemMixin, so now any object can reinstantiate itself
by calling self.reCreate(). The intent is really for the daboDesigner to be able 
to reinstantiate an object at will, such as when a property changes that won't
take effect at runtime, such as Alignment, BorderStyle, PasswordEntry, and all the
useful properties of dOptionGroup.