Refactor the API of dBizobj, making most of the 'properties' into actual Python properties complete with getters, setters, and docstrings. Here is a list of the new bizobj properties:
	Caption (new: friendly name for communicating with the user)
	DataSource (was dataSource)
	SQL (was sql)
	RequeryOnLoad (was noDataOnLoad)
	AutoPopulatePK (was autoPopulatePK)
	Parent (was _parent)
	KeyField (was keyField)
	LinkField (was linkField)
	ErrorMessage (was _errorMsg, and I think this one can go because of the exception-refactor)
	RequeryChildOnSave (was requeryChildOnSave)
	NewChildOnNew (was newChildOnNew)
	NewRecordOnNewParent (was newRecordOnNewParent)
	FillLinkFromParent (was fillLinkFromParent)
	RestorePositionOnRequery (was savePosOnRequery)

Obviously, I've FUBAR'd the original API so all the demos that use bizobj's 
will have to be changed, which I'll do after lunch.

Refactor getPropertyList() and getPropertyInfo() out of dPemMixin and into the
new dabo.common.PropertyHelperMixin, since these functions can be useful at a
lower level than our wx interface. Make dPemMixin and dBizobj inherit from
PropertyHelperMixin, which means that dBizobj can now theoretically be represented
on the designer and have its properties edited there.

Add beforeInit() hook to dBizobj, which allows the user code an opportunity to 
modify properties such as RequeryOnLoad before they take effect.

dBizobj.__setattr__ had old-style class code that was causing my new property API
to fail silently, and it drove me crazy for an entire hour at least. :)  Changing
to the new-style super() call for non-cursor field attributes fixed the problem.
Lesson learned: __getattr__ will be called only after the attribute was not found, 
while __setattr__ gets called before Python looks for the attribute at all. Sounds 
like a subtle difference on paper but the implications are huge!

Add method getChildren() to dBizobj, which returns a tuple of object references
to the children. 

Add icon childview.png, which represents a child bizobj in the runtime dFormDataNav.

In dFormDataNav, delay instantiation of the pageframe until the primary column def
has been assigned, which allows the childviews to appear in the nav menu and in the
page frame.

The UI for the child view stuff isn't working yet, but it is a little closer. I started
out this morning working on that and got sidetracked with the bizobj changes.