Removed the calls to requeryAllChildren() and afterPointerMove() from the bizobj. Added those calls as an option to the seek() method.
The first two methods are "for internal use only", and should never be called from
an app. They exist so that the bizobj can move the pointer to where it needs
to be without triggering a slew of additional effects. 

moveToPK() is only used in a requery() to try to restore the record pointer
to the PK of the record it was on before the cursror was requeried. It is
followed by a call to requeryAllChildren() and afterPointerMove() in that
method; calling it again from this one is just inefficient..

moveToRowNum() is called by the validate() method so that a bizobj can check
all of its records. You don't want it requerying its children, as they will
need to be validated separately, and will be called in turn.

In VFP Codebook, these methods are protected. For record pointer movement
from within an app, the public methods first(), next(), etc., should be used
to move the record pointer.

Since seek() is intended to be called both internally and from developer's
code, I've added a parameter 'runRequery' which defaults to False. If this
value is passed as True, requeryAllChildren() and afterPointerMove() will be
called if the cursor moves the pointer.