Handle NoRecordsError in dBizobj.getParentPk().
Childviews in dFormDataNav use dGrid in the childview pages just like in
the main browse page. Add logic so that when new/edit/delete is selected
in the popup menu, it gets directed to the proper place.

Raise NameError if user code tries to set a dObject's Name property to 
a sibling's name. If this happens inside the init phase of the control, the
error won't propagate up but the object will get assigned a new name 
instead.

Make control's init's call private methods _beforeInit() and _afterInit()
instead of the user-code beforeInit() and afterInit() hooks. This is to 
offer a level of guarantee that the basic stuff happens even if the user
forgets to doDefault(). _beforeInit() and _afterInit() now call the user-
code hook methods.

Add initChildObjects() hook, which is where daboDesigner should put its
addObject() code. Called from dPemMixin._afterInit().

Fix dSlider's docstring.