Add PageClass and PageCount properties to dPageFrame, for specifying the number of pages (new pages will be based on PageClass, allowing the user to override with their own base class). User still has the option to use AddPage() directly.
Reworked the setting of the default number of pages... does that work
for you, knowing that we aren't encouraging passing values to the
constructors? The only drawback to setting a positive number of pages
is that user code will have to first set to zero if they want to add
their own subclassed pages.

Add TabPosition to the list of properties that require recreation.