The user can use bizobj.validateRecord() in a couple ways: 	1) return errorMessageText, and dBizobj will raise the exception 	2) raise the exception itself. In case #2, None is returned, resulting in an exception in dBizobj. This code fixes that.
If a user were to requery, and then answer yes to save changes, but 
the save failed, the requery would happen anyway, wiping out all 
their changes. This code fixes that.

There are still some funky problems though. Do this:
	1) add a new record and save
	2) add a new record and make the bizrule fail.
Note that the bizrule error text is listed twice, and that the browse
screen has reverted the original successfully-saved record back to the
value it had when it was new.