Bugfix: refactor old pre-dException code in dBizobj.requeryAllChildren().
Tried to change default of dBizobj.RequeryOnLoad() to False instead 
of True, but look at that it was already set to False. I guess I 
slipped this in before and forgot. I'll fix the demos.
	
Recently, when we started forcing unique sibling names, I made the
constructor append the id of the control to make the name unique if
needed. I've improved this somewhat by making it an auto-incrementing
integer instead (txt1, txt2, txt3, etc.).

Add event handlers OnCreateWindow and OnDestroyWindow, and bind them 
to EVT_WINDOW_CREATE and EVT_WINDOW_DESTROY for all controls.

Add new dEvent, EVT_VALUECHANGED, and raise this event whenever a
control's value changes, whether programatically or interactively.

Add new function getAbsoluteName() to dPemMixin, which returns self's
fully-qualified name, such as 'dFormRecipes.dPageFrame.Page1.txtName'.

Remove the property list for the form being output to the shell, which
was causing problems on Windows and isn't necessary anyway.

Refactor some of the Select page in dFormDataNav, and make the select
checkboxes and textboxes persist their values to the user settings 
table. Now the select page will remember what you last searched on
and present it to you for convenience.

Actually, that persisting of values that I use in the select page is
now available for all data controls, by setting the new property
SaveRestoreValue.