Add property dApp to dPemMixin, a reference to the dApp application object when dApp is present or None. This gives convenient access to dApp from all controls that inherit from dPemMixin (almost all Dabo controls and forms).
Make dApp inherit from PropertyHelperMixin and DoDefaultMixin to conform
to most of the rest of the Dabo objects.

Add property AutoNegotiateUniqueNames to dApp, default True. When true,
if an object tries to name itself identical to a sibling, a unique name
will be found by appending an integer at the end. When false, a NameError
will be raised which was the previous behavior. 

The auto-negotiation of names is definitely implicit, and could potentially 
cause confusion when debugging problems. Consider:

	self.addObject(dabo.ui.dCheckBox, 'myCheckbox')
	self.addObject(dabo.ui.dCheckBox, 'myCheckbox')

No errors would have been raised, and two checkboxes got created, but 
the second checkbox got a name that differs from the name specified by the
programmer ('myCheckbox1' instead of 'myCheckbox').

I think the convenience outweighs the concerns about implicitness, but 
would be willing to change the default to False instead of True, which may
be better as the programmer would have to explicity tell dApp to act
implicitly, and presumably understand the consequences by doing so.