Modified the cursor to handle calculated and derived fields that are not to be included in updates to the backend database. The cursor mixin class has a user-editable list property called 'nonUpdateFields', and there is a private list property named '__nonUpdateFields' that is set by the cursor mixin after each requery. For most cases, the built-in behavior should suffice to prevent illegal update statements resulting from trying to update a derived field. The user-configurable property is there to allow the user to specify additonal fields to be ignored during an update.
The list of user-defined non-updatable fields is set by the bizobj's
'NonUpdateFields' property. Accessing the NonUpdateFields property will
return the concatenated list of the public and private non-update lists.

Also fixed a bunch of edge conditions dealing with the fact that until a
query is run, the cursor doesn't have many of the attributes that were
expected, such as _rows, rownumber, etc.