Rename dGrid to dGridDataNav, as it is really not meant as a general grid control but rather to fit in with the dFormDataNav subframework. It assumes that it is a member of a dPageDataNav, for instance. At some point, we'll want to define a basic dGrid which dGridDataNav will inherit, but the grid is one of the most complex UI controls so it'll have to wait for now.
Add new property, FormType, to dFormDataNav. There is 'Normal', which 
is a normal form with Select/Browse/Edit/n childview pages; 'PickList' 
which doesn't display the edit/childview pages and is intended to be
used to pick a record from a lookup; and 'Edit' which isn't implemented
yet but will only display the edit page(s) for the record with the 
given pk value.

Add new event, EVT_ITEMPICKED, for use by dFormDataNav when instantiated
as a picklist.

Make dGrid behave differently in several places, depending on whether 
its form is a picklist or a normal nav form.

Make the picklist form close upon a keypress of Escape. Make the toolbar/
navigation menu display only the appropriate items depending on FormType.

Define a childBehavior dict in dFormDataNav, which lets subclasses 
define certain things about how the childview page(s) will behave. Note
that this isn't how I want it to stay, this was just to get up and
running. Eventually, users will be able to easily subclass all of the
separate components of dFormDataNav, not merely the form itself, and
there will be appropriate hooks where they can put their code.

Update the childview page to add and delete child records. Editing child
records isn't implemented yet. Note that until the non-unique pk problem
is fixed as discussed this morning, weird things can happen when adding
child records, especially more than one of them, and it will fail if you
try to add child records to a newly-added parent record. 

--

(I also just tweaked my svn-repository commit-notify email script, to 
hopefully not give the diff of deleted files. So, if it worked, there
should not be a diff for the deleted dGrid.py.)