Introduce dSecurityManager, which handles validating a user login and caching group membership information. Introduce property SecurityManager to dApp object, allowing users to subclass dSecurityManager to do what they need it to do, and then attach an instance of their subclass to dApp.
Make dApp check with the SecurityManager if available, before starting the
main application event loop.

Introduce new UI class dDialog, which is a very simple frame usually meant
to be shown modally and to return a value. Introduce new UI class dLogin, 
which is a basic login form.

Wrap some property docstrings in _(). 

Start experimenting with Accelerator Tables, which can be used to workaround
the lack of Cancel and Default properties in command buttons.