Add a place for messages to be added to the login form, and set up dSecurityManager to pass on informative messages during the login process.
Add logged-in user name to the caption of the main form. Display 
the application name in the caption of the login form.

Move the ordering of app instantiation steps around somewhat so that
the login form displays before the main form. Now, you only get to 
see the main form after a successful login.

Base dDialog on dFormMixin, which lets it harness some useful form 
properties.