Change default __params value to the empty tuple in dBizobj. Since different backends expect different types for the params arg, modify dCursorMixin.execute to not send the params arg at all if it is None or otherwise empty. Tested with MySQL and Firebird: Ok.
Modified dbFirebird's getCursor() to do the required mixing-in and to return
a full Dabo cursor instance. dbFirebird now gets past all the initial errors
in my wiz-generated test app. Next up: there is MySQL-specific code in my
dPageDataNav code - dbFirebird is currently choking on the LIMIT clause.

Modified dProgressDialog to properly react to exceptions that happen in the
passed function. It used to be that if an exception happened in the requery()
code, the dProgressDialog wouldn't handle it and therefore the worker thread
would never return control back to the main thread. Now, control is returned
and the main thread tries to show useful information about the exception. It
is a start at least.