New controls: dDropdownList is an implementation of the wx.Choice control. It provides a means of presenting users with a non-editable dropdown list for selections.
dDateTextBox is a subclass of dTextBox that is specially modified to handle
dates. Currently it supports only two date formats (MM/DD/YYYY and
YYYY-MM-DD), but is easily modified to support whatever is needed. Users can
navigate to different dates by pressing shortcut keys. These keys are the
same as those used in Quicken, and should be familiar to many people.