I like the new dDateTextBox *much* better than my feeble attempt at a dDateControl, so I added a Deprecation warning to dDateControl, and made a couple improvements to dDateTextBox (at least I think they are improvements <g>). Also, changed the dPageDataNav edit page to use the new control instead of the old for date values.
The changes I made to dDateTextBox basically filter out the characters we know
we don't want and eat them. Also, I made it so the event doesn't need to be sent
to adjustDate() which simplifies it somewhat.

The next thing I see as necessary for dDateTextBox is basic validation and smartness,
such as entering '0622' and that getting expanded to '06-22-2004'. Also, more 
formats (duh) and moving the format specification into a real property that can
be set at runtime and defaults to a system or application property setting, such as
dabo.app.DateFormat = "MDY".

Another thing I'd like to see is the use of Python's datetime() instead of wx's.
This could potentially allow this class to be moved out of wx-specificness into 
the main ui directory (once other specifics, such as events, are properly abstracted.)

I'll delete dDateControl in a few weeks.