Added code to walk the sizer/control hierarchy and draw outlines for the various sizers. The behavior is off by default, but can be enabled by setting the form's DrawSizerOutlines property to True.
Besides being of interest at design time, it also served as a proof of
concept to me for creating a full XRC-like xml rendering of the form. I'm
looking at sticking to XRC, but I don't know if it will be able to handle
all of the Dabo differences, so a superset of XRC might be needed. We'll
see.