(FYI: This commit touches pretty much everything in the UI.)
Began work on abstracting the Events. dPemMixin now has three new methods:
bindEvent(), unBindEvent(), and raiseEvent(). These three events can be used
whether the event is based on a wx-event or is a dabo-custom event. Previously,
there was boilerplate code to raise an event and to bind a dabo custom event.
Now, they all look the same and just require standard one-liners.

The available events are listed in dabo.ui.dEventBase, and I tried to name them 
sanely and consistently. It is just a start: it only really represents events 
that we were already using in Dabo, excluding menu and grid events which I 
haven't tackled yet. I don't think that the event list will live long-term in 
dEventBase, as I believe they really belong in dPemMixin and, for the 
specialized events, in the individual controls themselves, much like we 
implemented the properties. In any case, we'll want to get a list of available 
events for a specific control for use with the IDE, so we'll have to think about
how to provide that.

Examples of binding an event to a method and raising an event:
	self.bindEvent(dEvents.MouseRightClick, self.onRightClick)
	self.raiseEvent(dEvents.RowNumChanged)
	
Changed some of the callback methods to reflect the new names of the events with
'on' prepended. In some cases, this meant changing 'On' to 'on', and in others
it meant changing the method names as well. I'm 99% confident that I changed all
necessary code to reflect these changes. <g>

Added more default event-bindings and moved that to dPemMixin: it was in 
dControlMixin. I'm still not sure if we want these default event bindings at
all, as the user should be able to easily bind the events to their own methods
as needed, but it was convenient for testing and there are probably good arguments
for leaving it as-is (we have code spattered around that takes advantage of the
auto-bindings).

Added all the current dControls to the test.py script inside ui/uiwx. This is
a convenient place to watch the events in action.

Fixed an ongoing problem with dTextBox: the value was being set to 'dTextBox' in
__init__, where it was intended to be setting the name.

Made dToggleButton into a data-aware control, as it can be used that way as well,
similar to VFP's graphical checkbox.

Made minor changes, such as changing wx calls to dabo calls (GetParent() to Parent, 
for example). Changed the default bitmap in dBitmapButton to a smaller dabo icon.

Removed the old, stupid, dDateControl as the new dDateTextControl is way better.