Back to the event handling. After realizing that Tkinter has no (obvious) way of letting me define and raise a custom event, and after having various issues with wx as well, I decided that the "right way" to go would be to design a Dabo event handler that would know how to bind and raise Dabo events, and the Dabo events would be defined in a UI-agnostic fashion.
It seems to work beautifully. Tested on wx and what I have so far of tk.

dEvents are now defined in the dabo.dEvents module, and are all based on
dabo.dEvents.Event. Users can define and raise their own events by simply
subclassing dEvents.Event.

The individual UI libs are responsible for catching ui-specific events and
then raising the Dabo events for user code to catch. Dabo automatically fills
in an EventData dictionary event property that contains the useful information
provided by the uiEvent (mouse and keyboard condition, for example), and it
is the Dabo event that gets passed to user callback functions.

The event handling is implemented in the new EventMixin class, and is 
mixed-in to dObject. This means that all Dabo objects, including cursors 
and bizobjs, could raise their own events as well (dCursor could raise
dEvents.RowNumChanged, for instance).

Don't worry: the interface for user code hasn't changed with this: it is 
still bindEvent(), raiseEvent() and unBindEvent().

--

Moved the LogEvents property to dObject, and made it smartly default to the
setting of its parent. Made dApp descend from dObject. All this makes it 
easy to log all events for all objects in the application by simply doing:
	app.LogEvents = ["All"]

---

Moved getAbsoluteName() from dPemMixin to dObject. ToDo: Make it a property
AbsoluteName.