If an object's Parent doesn't descend from dObject (there are some test scripts where this is the case) the getAbsoluteName() would fail. Gave these types of objects a name of "?".
I made a mistake moving dDataControlMixin to ui out of uiwx, as the Value property
is UI specific, and Value definitely belongs in dDataControlMixin. So I moved
dDataControlMixin (and, to be consistent, dControlMixin) back to uiwx but added
ui/dDataControlMixinBase and dControlMixinBase for the common code to move to.

Because of the recent moving of parts of dControlMixin to dControlMixinBase, the
UI lib must be set before defining any classes based on dControlMixin. This has 
actually been the case for weeks now, but we were able to get away with not doing
the dabo.ui.loadUI("wx") thing from inside the uiwx directory until now. Added 
conditional loading of wx into the ui from each of the dCheckBox, etc. scripts inside
uiwx so that the tests will run.

Ed, with these changes I was able to run wizMaintenanceApp and all the demos.