Added property Continue to dEvent, which lets event handlers specify whether to let additional handlers get a whack at the event, much like wx's Skip() function. The difference is that by default all Dabo events will continue - you only need to set Continue to False if you don't want further event handlers to run. This links up to the original ui's event as well, so if you don't want the default wx event handler to get the event, you just set evt.Continue = False.
Prior to this commit, there was no way to tell the default wx handler to
not run - iow there was no way to say "don't process the keypress" which
was a real limitation of the new dEvents model. 

Fixed the implementation of KeyChar: evt.EventData["keyChar"] now contains
the character entered.

Added MouseMove event, and implemented it for wx.

Added some code to dDateTextBox to reflect that the dEvents are now
working for a couple of the event bindings there. Ed: please test and let
me know. I set the useWxEvents flag to True and it appears to work.