Added dEvents.Paint, which fires when the control must redraw itself.
Made the internal _onWxEvent functions super-private.

Changed the LogEvents property to allow logging of all events except
a list of events to exclude from the log. For example, to log all 
events except for MouseMove and Idle, you can do:
	
	self.LogEvents = ["All", "MouseMove", "Idle"]

	
To log no events (the default):

	self.LogEvents = []
	
	
To log specific events:

	self.LogEvents = ["GotFocus", "LostFocus"]