Renamed the 'fieldSpecParser' to 'specParser', since it now handles more than just field spec XML files.
Updated the DataNav form to properly set relations based upon the new
relationSpecs.xml file created by the appWizard.

Added a method to the bizobj named 'addChildByRelationDict()' that accepts a
python dictionary (as returned by the specParser) and determines if there
are child bizobjs that need to be created. This will work recursively, so a
single call to the primary bizobj will establish all pertinent child
relationships.