Added a function to the lib directory named 'getMouseObject' that does in Dabo what SYS(1270) does in VFP: returns a reference to the object currently under the mouse pointer when the function is called.
I really wanted to put this in the main dabo directory, so that code using
it can be written simply as 'dabo.getMouseObject()', since most of the use
for this function will be in interactive mode. But it makes more sense to
keep it in the lib directory, since it is a library function.