Removed the test directory in biz.
Commented out all doDefault() calls and replaced with super().method() calls:
I ran a test and discovered that the latter executes 1.5 orders of magnitude
faster, and while we are talking 1/1000's of a second versus 1/100's of a 
second, if lots of objects are being created it could conceivably be noticeable.

I think that we should still show doDefault() in user documentation and demos,
but that we should try to use super(cls, self).method() in the dabo library
code. It is more explicit, doesn't really require that much more typing, and
more efficient at runtime. 

The standard super() will fail if the method doesn't exist in the superclass, while
doDefault() has code to pass silently in that case. So doDefault() is definitely
the appropriate method to show to users.