When a DataNav form was requeried, the nice little status bar message that displayed the results was immediately overwritten with the location of the cursor in the grid. I thought that the result was more important, so I changed some of the routines that display the status text to allow me to set the requery results last.
Also began adding code to allow the user to specify the OrderBy clause. Right now I'm using a right-click on the label, but that's not as user-friendly as I'd like. I'm thinking about moving the logic to a button that changes state as they click on it, but we'll see...